First working vec SystStructs seriliation, needed to add one for internal PRospec also
